backend/requirements.txt
aiosql==3.2.0
alembic==1.6.5
anyio==3.1.0
asgi-lifespan==1.0.1
asgiref==3.3.4
asyncpg==0.23.0
attrs==20.2.0
bcrypt==3.2.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==8.0.1
cloud-init==19.1.6
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.7
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.2
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
httpcore==0.13.4
httptools==0.2.0
httpx==0.18.1
idna==2.8
iniconfig==1.1.1
Jinja2==3.0.1
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
language-selector==0.1
loguru==0.5.3
Mako==1.1.4
MarkupSafe==2.0.1
netifaces==0.10.4
oauthlib==3.1.0
orjson==3.5.3
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyGObject==3.36.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-asyncio==0.15.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.18
SQLAlchemy-Utils==0.37.6
ssh-import-id==5.10
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ufw==0.36
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
